pyproject.toml
[project]
name = "parserlib"
version = "0.1.0"
description = "Downloader for manga and ranobe"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "fpdf2>=2.8.7",
    "msgspec>=0.20.0",
    "pyrate-limiter>=4.0.2",
    "requests>=2.32.5",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
minversion = "9.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
  "-vv",
  "--import-mode=importlib",
  "-m",
  "not api",
  "--cov=parserlib",
  "--cov-report=term-missing",
  "--cov-report=html"
]
markers = [
    "api: tests for checking external APIs"
]

[tool.coverage.report]
show_missing = true
skip_covered = false
